Remove 'coefficients' methods
From help(coef):

"All object classes which are returned by model fitting functions should
provide a ‘coef’ method or use the default one.  (Note that the method
is for ‘coef’ and not ‘coefficients’.)"

i.e. stats::coefficients calls the coef method anyway, so we don't need
to do anything about that.